Fix algorithm, trying to see if the new way of calculating work,

+++ app/src/main/java/Logic/Velocity.java
@@ -15,7 +15,7 @@ public class Velocity {
         z = _z;
     }
 
-    public static Velocity getVelocity(double acceX, double acceY, double acceZ, double time) {
+    public static Velocity getVelocity(double acceX, double acceY, double acceZ, long time) {
         return new Velocity(acceX * time/1000, acceY * time/1000, acceZ * time/1000);
     }
